+++ lib/writer.js
@@ -1,11 +1,13 @@
 var alru = require('array-lru')
 var hash = require('./hash')
-var encoding = require('./encoding')
+var trie = require('./trie')
+var messages = require('./messages')
 
 module.exports = Writer
 
 function Writer (feed, id, codec) {
   if (!(this instanceof Writer)) return new Writer(feed, id, codec)
+
   this.id = id
   this.feed = feed
   this.cache = alru(8192)
@@ -29,14 +31,23 @@ Writer.prototype.head = function (cb) {
 }
 
 Writer.prototype.append = function (node, cb) {
-  this.feed.append(encoding.encode(node, this.codec), cb)
+  var buf = messages.Node.encode({
+    key: node.key,
+    value: this.codec ? this.codec.encode(node.value) : node.value,
+    heads: node.heads,
+    trie: trie.encode(node.trie)
+  })
+
+  this.feed.append(buf, cb)
 }
 
 Writer.prototype._decode = function (val, seq) {
-  var node = encoding.decode(val, this.codec)
+  var node = messages.Node.decode(val)
   node.feed = this.id
   node.seq = seq
   node.path = hash(node.key, true)
+  node.trie = trie.decode(node.trie)
+  if (this.codec) node.value = this.codec.decode(node.value)
   this.cache.set(seq, node)
   return node
 }
